Performance improvements to scrolling hexdump
Instead of reading and printing each byte, we read+push 8 bytes to the stack and pull+print afterwards. This allows us to minimise the number of times interrupts & ROM have to be toggled.

Printing a byte as hexadecimal was sped up as X & Y are already preserved by CHROUT.

Added VSCode settings to the project.